+++ src/Trade360SDK.Common.Entities/Entities/OutrightLeague/OutrightLeagueCompetitionWrapper.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 
 namespace Trade360SDK.Common.Entities.OutrightLeague
 {
@@ -7,6 +8,7 @@ namespace Trade360SDK.Common.Entities.OutrightLeague
         public int Id { get; set; }
         public string? Name { get; set; }
         public int Type { get; set; }
+        public DateTime? NextFixtureStartTime { get; set; }
         public IEnumerable<OutrightLeagueEventsWrapper<TEvent>>? Competitions { get; set; }
     }
 }